# backend/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Internationalized culture / e-commerce / custom-travel platform for Jingdezhen ceramic art.\nBackend in Go + Fiber; money as BIGINT minor units (CNY base, USD/EUR/GBP presentment).",
        "title": "Jingdezhen Ceramics Platform API",
        "termsOfService": "https://jingdezhen.test/terms",
        "contact": {
            "name": "Jingdezhen Ceramics Platform Team"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/webhooks/airwallex": {
            "post": {
                "description": "Receives Airwallex payment events. Verifies the signature via\nthe gateway, then acks 200 immediately (enqueue-and-ack). A\nsignature mismatch returns 400. Idempotent on the idempotency_key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks",
                    "payments"
                ],
                "summary": "Airwallex webhook",
                "responses": {
                    "200": {
                        "description": "OK (event acknowledged + enqueued)"
                    },
                    "400": {
                        "description": "Webhook signature verification failed",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Payment gateway not configured",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhooks/paypal": {
            "post": {
                "description": "Receives PayPal payment events. Verifies the signature via the\ngateway, then acks 200 immediately (enqueue-and-ack). A signature\nmismatch returns 400. Idempotent on the idempotency_key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks",
                    "payments"
                ],
                "summary": "PayPal webhook",
                "responses": {
                    "200": {
                        "description": "OK (event acknowledged + enqueued)"
                    },
                    "400": {
                        "description": "Webhook signature verification failed",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Payment gateway not configured",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wishlist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginated list of the signed-in user's favorited SKUs, enriched\nwith the parent product's display info for the requested locale.\nLocale resolution: ?locale= overrides Accept-Language.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "List the current user's wishlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"en-US\"",
                        "description": "BCP 47 locale (e.g. en-US). Overrides Accept-Language.",
                        "name": "locale",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size (max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.WishlistItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Favorites a SKU. Idempotent: favoriting an already-favorited\nSKU is a no-op (still returns 201). A nonexistent SKU returns 404.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Add a SKU to the wishlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "SKU to favorite",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.AddWishlistRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created (empty body)"
                    },
                    "400": {
                        "description": "Invalid body / validation",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "SKU not found",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wishlist/{sku_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Unfavorites a SKU. Removing a SKU not in the wishlist returns 404.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Remove a SKU from the wishlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "SKU ID",
                        "name": "sku_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content (empty body)"
                    },
                    "400": {
                        "description": "Invalid SKU ID",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Wishlist item not found",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "jingdezhen-ceramics-backend_internal_models.AddWishlistRequest": {
            "type": "object",
            "required": [
                "sku_id"
            ],
            "properties": {
                "sku_id": {
                    "type": "integer"
                }
            }
        },
...
        "jingdezhen-ceramics-backend_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Optional additional details",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "jingdezhen-ceramics-backend_internal_models.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "jingdezhen-ceramics-backend_internal_models.WishlistItem": {
            "type": "object",
            "properties": {
                "artist_name": {
                    "type": "string"
                },
                "attributes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "favorited_at": {
                    "type": "string"
                },
                "price_cny": {
                    "description": "minor units (fen)",
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_slug": {
                    "description": "translation",
                    "type": "string"
                },
                "product_title": {
                    "description": "translation",
                    "type": "string"
                },
                "sku_code": {
                    "type": "string"
                },
                "sku_id": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                },
                "thumbnail_url": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003caccess_token\u003e\" — JWT access token from POST /auth/login (or OAuth). Required for all /admin/* and customer-authed endpoints; omitted for public /catalog/* reads.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}